Now with markup!
A post has two body fields: body and body_markup. Body is edited by the user himself. The post may be marked up with a markup language. django-markup is being used to specify which markup language the user wants (markdown is default), and to actually render the markup. The rendered markup is stored in body_markup, which is the field that is actually printed by the templates. This is more resource efficient than using django.contrib.markup to render the markup in the template each time the post is viewed.

The code for this was copied from montylounge's (https://github.com/montylounge) fork of django-basic-apps.