Ignoring migrations
I'm not sure what the best way is to handle migrations in a reusable app. With them in there, if someone downloads this project and attempts a fresh install, they won't be able to do a syncdb because South will try to handle it. But if they try to migrate, that will also fail because none of the database tables exist yet.

So, for now, I'm ignoring migrations in the rep. Someone doing a fresh install can just do a normal syncdb. The next time I change the models around, it's easy enough for someone following the project to have South generate their own schemamigration.