Tease can now be marked up.
I went back and forth on whether or not to implement this. The tease field should be kept simple. But I find myself wanting to put a link or a paragraph in the tease content now and again, and I'd rather to that with Markdown than HTML. For now, it's in there. I may remove this in the future.