Adding disqus comment identifier.
I removed this in an earlier commit because it caused some comments to not show up on posts. I thought the problem was somwhere in django-disqus' export function. Now I've had time to look into it, and I found that I was right. Here's the skinny:

When comments are exported, first it checks to see if the comment's thread already exists. If it does not, it creates a new thread. When it creates a thread, that thread's idenitifer is set to whatever is returned by the comment object's site_object function. That defaults to the post object's unicode function, which I have set to the post's title. So the thread's identifier is the title of the post. Obviously things won't work if in the template you tell Disqus that the identifier is the post's ID.

I think using the ID makes more sense than the title (titles may change), so that's what I'm putting in the template. I deleted all my comments on Disqus and re-exported them, but first I went into my models and changed the post object's unicode function to return self.id. After the export was complete, I changed that back to self.name. Everything should be working dandy now.